refactor(api): PR-6 commit 4 — system router 11 routes (updater 域)
server.py 3548 → 3194 行（−354），11 system routes + 1 BaseModel +
重启 / SIGINT / no-running-task 三个 helper 全部抽到 api/routers/system.py。

  api/routers/system.py  11 routes  restart / version / update_check / update /
                                    rollback / update_status / update_log /
                                    preflight / dev_commits / init_git /
                                    release_notes
                                    + _raise_sigint_after_response + _check_no_running_tasks
                                    + _RESTART_FLAG / _SHUTDOWN_FORCE_EXIT_TIMEOUT
  api/schemas/system.py  1 BaseModel UpdateRequest

server.py 顶部清理（不再被引用的 import）：
  - updater / release_notes_svc（搬到 system router）
  - flash_attention_setup / onnxruntime_setup / pending_install / torch_setup /
    xformers_setup（搬到 installs router）
  - system_stats（搬到 health router 走自己 import）

测试跟进：tests/test_studio_server.py 把 patch path 从
`server._raise_sigint_after_response / server._RESTART_FLAG` 改到
`studio.api.routers.system._raise_sigint_after_response / _RESTART_FLAG`
（PR-5 PR description 写的 lesson 第二次应用 — fixture monkeypatch path 跟
搬迁）。

回归：PR-1 安全网 152 + studio_server 44 + presets/import smoke = **227/227 全绿**。
route snapshot 不变（160 routes）。
