+++ studio/api/schemas/system.py
@@ -1,0 +1,8 @@
+"""/api/system/update 请求 BaseModel（PR-6 commit 4 从 server.py 抽出）。"""
+from __future__ import annotations
+
+from pydantic import BaseModel
+
+
+class UpdateRequest(BaseModel):
+    target: str = "origin/master"  # ref / commit sha / origin/branch
